Remove limits.hpp and invalid static_asserts
The static_asserts relate to a old and wrong assumption that code size
cannot exceed 0x6000 (this is only true for deployed code but not
initcode). The overflow problem is already solved in advanced analysis
and these checks are simply misleading. They also use all 3 constants
defined in limits.hpp so the file can be removed.