precompiles: Increase output buffer size (#695)
Previously we were passing the max output size computed in analysis. This was working only in case the computed max output size was not bigger than the allocated buffer. In case it was bigger: assert failure or stack buffer overflow.

Now we just pass the allocated buffer size. Similarly, if the allocated buffer is big enough all works correctly. If not we will get the similar assert failure or the output will be trimmed to the allocated size (therefore incorrect).

None of this is final solution. We have to switch to a dynamic buffer allocation or eliminate the intermediate buffer.