baseline: VM-level execution states
This adds an array of `ExecutionState` objects to the `VM` instance:
one lazily initialized object for all possible call depths.
This also makes the `VM` non-thread safe, but it is fine for current
users. We can improve on this in the future by spiting the the VM into
two parts: non-thread safe opaque execution context
and a thread-safe VM with user configuration.